+++ app/src/main/java/woowacourse/movie/view/mapper/ReservationMapper.kt
@@ -1,0 +1,12 @@
+package woowacourse.movie.view.mapper
+
+import com.example.domain.Reservation
+import woowacourse.movie.view.model.ReservationUiModel
+
+fun Reservation.toUiModel(): ReservationUiModel = ReservationUiModel(
+    movieTitle,
+    screeningDateTime,
+    seats.size,
+    seats.map { it.toUi() },
+    finalReservationFee.amount
+)
