+++ app/src/main/java/woowacourse/movie/view/model/ReservationUiModel.kt
@@ -1,17 +1,8 @@
 package woowacourse.movie.view.model
 
 import android.os.Parcelable
-import com.example.domain.Reservation
 import kotlinx.parcelize.Parcelize
 import java.time.LocalDateTime
-
-fun Reservation.toUiModel(): ReservationUiModel = ReservationUiModel(
-    movieTitle,
-    screeningDateTime,
-    seats.size,
-    seats.map { it.toUi() },
-    finalReservationFee.amount
-)
 
 @Parcelize
 data class ReservationUiModel(
